+++ aim/web/ui/src/pages/ImagesExplore/ImagesExplore.tsx
@@ -18,7 +18,9 @@ import IllustrationBlock from 'components/IllustrationBlock/IllustrationBlock';
 import pageTitlesEnum from 'config/pageTitles/pageTitles';
 import { ResizeModeEnum } from 'config/enums/tableEnums';
 import { RowHeightSize } from 'config/table/tableConfigs';
-import GroupingPopovers from 'config/grouping/GroupingPopovers';
+import GroupingPopovers, {
+  GroupNameEnum,
+} from 'config/grouping/GroupingPopovers';
 import { RequestStatusEnum } from 'config/enums/requestStatusEnum';
 import {
   IllustrationsEnum,
@@ -39,6 +41,7 @@ import { AppNameEnum } from 'services/models/explorer';
 
 import { IGroupingSelectOption } from 'types/services/models/metrics/metricsAppModel';
 import { IApiRequest } from 'types/services/services';
+import { IModel, State } from 'types/services/models/model';
 
 import exceptionHandler from 'utils/app/exceptionHandler';
 import getStateFromUrl from 'utils/getStateFromUrl';
@@ -52,7 +55,7 @@ import './ImagesExplore.scss';
 function ImagesExplore(): React.FunctionComponentElement<React.ReactNode> {
   const route = useRouteMatch<any>();
   const history = useHistory();
-  const imagesExploreData = useModel<Partial<any>>(imagesExploreAppModel);
+  const imagesExploreData = useModel<any>(imagesExploreAppModel);
   const imagesWrapperRef = React.useRef<any>(null);
   const tableElemRef = React.useRef<HTMLDivElement>(null);
   const wrapperElemRef = React.useRef<HTMLDivElement>(null);
@@ -167,13 +170,19 @@ function ImagesExplore(): React.FunctionComponentElement<React.ReactNode> {
       );
       appRequestRef
         .call((detail: any) => {
-          exceptionHandler({ detail, model: imagesExploreAppModel });
+          exceptionHandler({
+            detail,
+            model: imagesExploreAppModel as IModel<State>,
+          });
         })
         .then(() => {
           imagesExploreAppModel.setDefaultAppConfigData(false);
           imagesRequestRef = imagesExploreAppModel.getImagesData();
           imagesRequestRef.call((detail: any) => {
-            exceptionHandler({ detail, model: imagesExploreAppModel });
+            exceptionHandler({
+              detail,
+              model: imagesExploreAppModel,
+            });
           });
         });
     } else {
@@ -247,7 +256,7 @@ function ImagesExplore(): React.FunctionComponentElement<React.ReactNode> {
               />
               <Grouping
                 groupingPopovers={GroupingPopovers.filter(
-                  (g) => g.groupName === 'row',
+                  (g) => g.groupName === GroupNameEnum.ROW,
                 )}
                 isDisabled={isProgressBarVisible}
                 groupingData={imagesExploreData?.config?.grouping}
@@ -325,7 +334,7 @@ function ImagesExplore(): React.FunctionComponentElement<React.ReactNode> {
                         focusedState={
                           imagesExploreData?.config?.images?.focusedState!
                         }
-                        tooltip={imagesExploreData?.config?.images?.tooltip!}
+                        tooltip={imagesExploreData?.tooltip!}
                         sortFieldsDict={memoizedImagesSortFields.sortFieldsDict}
                         sortFields={memoizedImagesSortFields.sortFields}
                         additionalProperties={
@@ -349,9 +358,7 @@ function ImagesExplore(): React.FunctionComponentElement<React.ReactNode> {
                             selectOptions={
                               imagesExploreData?.groupingSelectOptions!
                             }
-                            tooltip={
-                              imagesExploreData?.config?.images?.tooltip!
-                            }
+                            tooltip={imagesExploreData?.tooltip!}
                             orderedMap={imagesExploreData?.orderedMap}
                             additionalProperties={
                               imagesExploreData?.config?.images