+++ aim/web/ui/src/types/pages/metrics/components/Controls/Controls.d.ts
@@ -11,7 +11,7 @@ import { IAxesScaleState } from 'types/components/AxesScalePopover/AxesScalePopo
 import {
   IAggregationConfig,
   IAlignmentConfig,
-  IPanelTooltip,
+  ITooltip,
   IChartZoom,
   IGroupingSelectOption,
 } from 'types/services/models/metrics/metricsAppModel';
@@ -25,7 +25,7 @@ export interface IControlProps {
   chartType: ChartTypeEnum;
   data: ILine[][] | any;
   selectOptions: IGroupingSelectOption[];
-  tooltip: IPanelTooltip;
+  tooltip?: ITooltip;
   ignoreOutliers: boolean;
   zoom?: IChartZoom;
   highlightMode: HighlightEnum;
@@ -37,7 +37,7 @@ export interface IControlProps {
   alignmentConfig: IAlignmentConfig;
   densityType: DensityOptions;
   selectFormOptions: ISelectOption[];
-  onChangeTooltip: (tooltip: Partial<IPanelTooltip>) => void;
+  onChangeTooltip: (tooltip: Partial<ITooltip>) => void;
   onIgnoreOutliersChange: () => void;
   onHighlightModeChange: (mode: number) => void;
   onDensityTypeChange: (type: DensityOptions) => void;