+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Enhancements:
 
+- Filter redundant tooltip data from URL config state (KaroMourad)
 - Improve rendering performance by enhancing table columns virtualization mechanism (roubkar)
 - Increase visibility and usability of the Show table diff button (arsengit)
 - Add support for tensorboard audios conversion (mihran113)
@@ -15,6 +16,7 @@
 - Fix tensorboard log conversion for images (mihran113)
 - Fix `SyntaxError` handling for python3.10+ (mihran113)
 - Check if gradient is None when tracking gradient distributions (kage08)
+
 
 ## 3.11.1 Jun 27, 2022